Fix isPartiallyVisible was returns false when child is larger than parent
If the child view is larger then parent container and child's start and end is out of bounds,
The ViewBoundsHelper.isPartiallyVisible was returns false. It is incorrect.
In this situation, the function should return true because event if child view start and end is
invisible, the child view is partially visible.
This commit fix this bug.